Jeigu vartotojo nuotrauka default.png jos neištrina

+++ backend/app/Http/Controllers/ApiController.php
@@ -35,7 +35,9 @@ class ApiController extends Controller
                 return response()->json(["error" => $validation->errors()]);
             } else {
                 $file = User::select('foto')->where('id', '=', auth()->user()->id)->get(); //Randa sena filePath
-                File::delete('../storage/app/public/' . $file[0]['foto']); //Keliant naują foto seną ištrina
+                if ($file[0]['foto'] != 'userimg/default.png') {
+                    File::delete('../storage/app/public/' . $file[0]['foto']); //Keliant naują foto seną ištrina
+                }
                 $path = $request->file('file')->store('public/userimg');
                 $filename = str_replace('public/', "", $path);
                 User::where('id', auth()->user()->id)->update(['foto' => $filename]);
